02-vector-db-bench/requirements.txt
numpy
pandas
tqdm
scikit-learn
faiss-cpu
sentence-transformers
typer
chromadb